# Replace HashMaps with a bit-vector for unique depth computation