--- a/flatgfa/Cargo.toml
+++ b/flatgfa/Cargo.toml
@@ -10,6 +10,7 @@ path = "src/cli/main.rs"
 [dependencies]
 argh = "0.1.12"
 atoi = "2.0.0"
+bit-vec = "0.8.0"
 bstr = "1.10.0"
 memchr = "2.7.4"
 memmap = "0.7.0"