--- a/flatgfa/src/ops/depth.rs
+++ b/flatgfa/src/ops/depth.rs
@@ -1,22 +1,20 @@
 use crate::flatgfa;
-use std::collections::HashSet;
 
 pub fn depth(gfa: &flatgfa::FlatGFA) -> (Vec<usize>, Vec<usize>) {
-    let mut depths = vec![0; gfa.segs.len()];
-    // Initialize uniq_paths
-    let mut uniq_paths = Vec::<HashSet<usize>>::new();
-    uniq_paths.resize(gfa.segs.len(), HashSet::new());
-    // do not assume that each handle in `gfa.steps()` is unique
-    for (idx, path) in gfa.paths.all().iter().enumerate() {
+    let mut depths = vec![0; gfa.segs.len()]; // The number of paths that traverse each segment.
+    let mut uniq_depths = vec![0; gfa.segs.len()]; // The number of distinct paths that traverse them.
+
+    for path in gfa.paths.all().iter() {
+        let mut seen = vec![0; gfa.segs.len()]; // Has this path traversed this segment?
         for step in &gfa.steps[path.steps] {
             let seg_id = step.segment().index();
-            // Increment depths
             depths[seg_id] += 1;
-            // Update uniq_paths
-            uniq_paths[seg_id].insert(idx);
+            if seen[seg_id] == 0 {
+                uniq_depths[seg_id] += 1;
+                seen[seg_id] = 1;
+            }
         }
     }
 
-    let uniq_depths = uniq_paths.iter().map(|set| set.len()).collect();
     (depths, uniq_depths)
 }