Just a low-hanging fruit optimization for this op: instead of a `Vec<HashMap<usize>>`, we can use a single `BitVec`.

I did this in three steps and measured their performance on the chr22 GFA:

* Switching to a dense vector, but not a proper bit vector: 14.7 -> 2.2 seconds
* Switching to a proper bit vector from the [bit-vec](https://docs.rs/bit-vec/latest/bit_vec/) crate: 2.2 -> 2.0 seconds
* Reusing a single bit vector instead of allocating in the loop: no effect; maybe the optimizer already got this

Anyway, a quick profiling run suggests that running `fgfa depth > /dev/null` on this graph is now 90% printing stuff and 10% actual depth computation, so I think we'll call that good.